refactoring and clean up
